+++ Source/Boilerplate.Templates/Content/ApiTemplate/Startup.cs
@@ -1,5 +1,6 @@
 ﻿namespace ApiTemplate
 {
+    using System;
 #if (Versioning)
     using System.Linq;
 #endif
@@ -31,7 +32,7 @@
     /// <summary>
     /// The main start-up class for the application.
     /// </summary>
-    public partial class Startup
+    public class Startup : IStartup
     {
         /// <summary>
         /// Gets or sets the application configuration, where key value pair settings are stored. See
@@ -129,7 +130,7 @@
         /// called by the ASP.NET runtime. See
         /// http://blogs.msdn.com/b/webdev/archive/2014/06/17/dependency-injection-in-asp-net-vnext.aspx
         /// </summary>
-        public void ConfigureServices(IServiceCollection services) =>
+        public IServiceProvider ConfigureServices(IServiceCollection services) =>
             services
 #if (ApplicationInsights)
                 // Add Azure Application Insights data collection services to the services container.
@@ -181,7 +182,8 @@
                 .AddCommands()
                 .AddRepositories()
                 .AddServices()
-                .AddTranslators();
+                .AddTranslators()
+                .BuildServiceProvider();
 
         /// <summary>
         /// Configures the application and HTTP request pipeline. Configure is called after ConfigureServices is